feat: Refactor the RightPanel Collapsable component and make it generic
