+++ src/components/Collapsable/Collapsable.types.ts
@@ -1,0 +1,8 @@
+export type Props = {
+  label: string | React.ReactNode
+  children: React.ReactNode
+}
+
+export type State = {
+  isCollapsed: boolean
+}
